chore: merge service ad handler into regular ad handler
