+++ pkg/clusteragent/instrumentation/handlers/autodiscovery.go
@@ -11,10 +11,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"hash/fnv"
-	"sort"
 	"strings"
-	"sync"
 
 	datadoghq "github.com/DataDog/datadog-operator/api/datadoghq/v1alpha1"
 	autoscalingv2 "k8s.io/api/autoscaling/v2"
@@ -28,40 +25,22 @@ import (
 
 const (
 	checksReadyConditionType = "ChecksReady"
-
-	// autodiscoveryProvider is the integration.Config Source value used for configs
-	// translated from a DatadogInstrumentation CR by the Autodiscovery handler.
-	autodiscoveryProvider = "datadoginstrumentation"
+	autodiscoveryProvider    = "datadoginstrumentation"
 )
 
-type CheckStore struct {
-	mu      sync.RWMutex
-	configs map[string][]integration.Config
-	// states maps namespace/name → "uid:generation". Including the UID ensures that a
-	// delete+recreate of a CR with the same name is detected even when the new CR
-	// starts at generation 1.
-	states     map[string]string
-	configHash uint64
-}
-
-func NewCheckStore() *CheckStore {
-	return &CheckStore{
-		configs:    make(map[string][]integration.Config),
-		states:     make(map[string]string),
-		configHash: fnv.New64a().Sum64(),
-	}
-}
-
 // AutodiscoveryHandler translates DatadogInstrumentation check sections into
-// integration.Config entries and stores them in memory for delivery to node agents.
+// integration.Config entries. It supports both workload targets (Deployment,
+// DaemonSet, etc.) and Service targets, branching internally on the target kind.
 type AutodiscoveryHandler struct {
-	checkStore *CheckStore
+	checkStore    *CheckStore
+	templateStore *ServiceCheckTemplateStore
 }
 
 // NewAutodiscoveryHandler returns the Autodiscovery DatadogInstrumentation handler.
 func NewAutodiscoveryHandler(dep *Deps) *AutodiscoveryHandler {
 	return &AutodiscoveryHandler{
-		checkStore: dep.CheckStore,
+		checkStore:    dep.CheckStore,
+		templateStore: dep.ServiceCheckTemplateStore,
 	}
 }
 
@@ -78,7 +57,7 @@ func (h *AutodiscoveryHandler) HasSection(cr *datadoghq.DatadogInstrumentation) 
 // SupportsTarget returns whether Autodiscovery check delivery supports the target kind.
 func (h *AutodiscoveryHandler) SupportsTarget(ref autoscalingv2.CrossVersionObjectReference) bool {
 	switch ref.Kind {
-	case "Deployment", "DaemonSet", "StatefulSet", "CronJob", "Job":
+	case "Deployment", "DaemonSet", "StatefulSet", "CronJob", "Job", "Service":
 		return true
 	default:
 		return false
@@ -90,8 +69,9 @@ func (h *AutodiscoveryHandler) Validate(cr *datadoghq.DatadogInstrumentation) []
 	return validateChecks(cr, h.Name())
 }
 
-// Handle translates check configs into integration.Config entries on Create/Update,
-// removes them on Delete, and reports a ChecksReady status.
+// Handle translates check configs on Create/Update, removes them on Delete,
+// and reports a ChecksReady status. For Service targets the configs are stored
+// as templates; for workload targets they are stored directly.
 func (h *AutodiscoveryHandler) Handle(_ context.Context, event instrumentation.EventType, cr *datadoghq.DatadogInstrumentation) (instrumentation.HandlerStatus, error) {
 	if cr == nil {
 		return instrumentation.HandlerStatus{
@@ -105,7 +85,12 @@ func (h *AutodiscoveryHandler) Handle(_ context.Context, event instrumentation.E
 	key := cr.Namespace + "/" + cr.Name
 
 	if event == instrumentation.EventDelete {
-		h.checkStore.deleteConfigs(key)
+		if isService(cr) {
+			h.templateStore.deleteTemplates(key)
+		} else {
+			h.checkStore.deleteConfigs(key)
+		}
+
 		return instrumentation.HandlerStatus{
 			Type:    checksReadyConditionType,
 			Status:  metav1.ConditionTrue,
@@ -116,7 +101,13 @@ func (h *AutodiscoveryHandler) Handle(_ context.Context, event instrumentation.E
 
 	configs := make([]integration.Config, 0, len(cr.Spec.Config.Checks))
 	for _, check := range cr.Spec.Config.Checks {
-		cfg, err := translateCheck(cr, check)
+		var cfg integration.Config
+		var err error
+		if isService(cr) {
+			cfg, err = translateServiceCheck(cr, check)
+		} else {
+			cfg, err = translateWorkloadCheck(cr, check)
+		}
 		if err != nil {
 			return instrumentation.HandlerStatus{
 				Type:    checksReadyConditionType,
@@ -128,7 +119,11 @@ func (h *AutodiscoveryHandler) Handle(_ context.Context, event instrumentation.E
 		configs = append(configs, cfg)
 	}
 
-	h.checkStore.setConfigs(key, configs, cr.Generation, string(cr.UID))
+	if isService(cr) {
+		h.templateStore.writeTemplates(key, cr, configs)
+	} else {
+		h.checkStore.writeConfigs(key, cr, configs)
+	}
 
 	return instrumentation.HandlerStatus{
 		Type:    checksReadyConditionType,
@@ -138,68 +133,18 @@ func (h *AutodiscoveryHandler) Handle(_ context.Context, event instrumentation.E
 	}, nil
 }
 
-func (c *CheckStore) ListConfigs() ([]integration.Config, uint64) {
-	c.mu.RLock()
-	defer c.mu.RUnlock()
-	out := make([]integration.Config, 0)
-	for _, cfgs := range c.configs {
-		out = append(out, cfgs...)
-	}
-	return out, c.configHash
+// ListConfigs returns a snapshot of all stored integration.Config entries
+// across all DatadogInstrumentation CRs handled by this instance and the
+// current state hash.
+func (h *AutodiscoveryHandler) ListConfigs() ([]integration.Config, uint64) {
+	return h.checkStore.ListConfigs()
 }
 
-// ConfigHash returns a deterministic hash of the current set of (key, generation) pairs,
-// consistent across all cluster agent replicas for the same CR state.
-func (c *CheckStore) Hash() uint64 {
-	c.mu.RLock()
-	defer c.mu.RUnlock()
-	return c.configHash
-}
-
-func (c *CheckStore) setConfigs(key string, configs []integration.Config, generation int64, uid string) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	if len(configs) == 0 {
-		delete(c.configs, key)
-		delete(c.states, key)
-	} else {
-		c.configs[key] = configs
-		c.states[key] = fmt.Sprintf("%s:%d", uid, generation)
-	}
-	c.configHash = c.hashStates()
-}
-
-func (c *CheckStore) deleteConfigs(key string) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	delete(c.configs, key)
-	delete(c.states, key)
-	c.configHash = c.hashStates()
-}
-
-// hashStates computes a deterministic hash from the sorted set of
-// "key:uid:generation" entries in the store. Including the UID ensures that a
-// recreation of a CR with the same namespace/name is detected even when
-// the new CR starts at generation 1.
-func (c *CheckStore) hashStates() uint64 {
-	keys := make([]string, 0, len(c.states))
-	for k := range c.states {
-		keys = append(keys, k)
-	}
-	sort.Strings(keys)
-	h := fnv.New64a()
-	for _, k := range keys {
-		fmt.Fprintf(h, "%s:%s\n", k, c.states[k]) //nolint:errcheck
-	}
-	return h.Sum64()
-}
-
-func translateCheck(cr *datadoghq.DatadogInstrumentation, check datadoghq.DatadogInstrumentationCheckConfig) (integration.Config, error) {
+func translateWorkloadCheck(cr *datadoghq.DatadogInstrumentation, check datadoghq.DatadogInstrumentationCheckConfig) (integration.Config, error) {
 	initConfig, instances, logsConfig, err := translateConfigChecks(check)
 	if err != nil {
 		return integration.Config{}, err
 	}
-
 	return integration.Config{
 		Name:          check.Integration,
 		ADIdentifiers: check.ContainerImage,
@@ -208,6 +153,20 @@ func translateCheck(cr *datadoghq.DatadogInstrumentation, check datadoghq.Datado
 		LogsConfig:    logsConfig,
 		CELSelector:   buildCELSelector(cr.Spec.TargetRef, cr.Namespace),
 		Source:        fmt.Sprintf("%s:%s/%s", autodiscoveryProvider, cr.Namespace, cr.Name),
+	}, nil
+}
+
+func translateServiceCheck(cr *datadoghq.DatadogInstrumentation, check datadoghq.DatadogInstrumentationCheckConfig) (integration.Config, error) {
+	initConfig, instances, logsConfig, err := translateConfigChecks(check)
+	if err != nil {
+		return integration.Config{}, err
+	}
+	return integration.Config{
+		Name:       check.Integration,
+		InitConfig: initConfig,
+		Instances:  instances,
+		LogsConfig: logsConfig,
+		Source:     fmt.Sprintf("%s:%s/%s", autodiscoveryProvider, cr.Namespace, cr.Name),
 	}, nil
 }
 
@@ -272,7 +231,6 @@ func buildCELSelector(ref autoscalingv2.CrossVersionObjectReference, namespace s
 }
 
 // validateChecks validates the check configs in a DatadogInstrumentation CR.
-// It is shared by all handlers that support the spec.config.checks section.
 func validateChecks(cr *datadoghq.DatadogInstrumentation, handlerName string) []instrumentation.ValidationError {
 	if cr == nil {
 		return nil
@@ -297,7 +255,8 @@ func validateChecks(cr *datadoghq.DatadogInstrumentation, handlerName string) []
 				HandlerName: handlerName,
 			})
 		}
-		if len(check.ContainerImage) == 0 {
+
+		if len(check.ContainerImage) == 0 && !isService(cr) {
 			errs = append(errs, instrumentation.ValidationError{
 				Type:        checksReadyConditionType,
 				Reason:      "InvalidContainerImage",
